-- ADDED DATABASE SETUP OF FOOD SCREEN -- FETCHING DATA BASED ON RECENT PRODUCT -- ADDED UPSERT OPERATION -- COMPLETED DELETE OPERATION OF FOOD ORDER -- MADE CHANGES IN HOME SCREEN    -- ADDED text "NO DATA FOUND"

+++ FoodGrid/app/src/main/java/com/example/foodgrid/dao/UserDAO.java
@@ -29,7 +29,7 @@ public interface UserDAO {
 
     // update userStatus
     @Query("UPDATE users SET userStatus = :updatedUserStatus WHERE userId = :userId")
-    public void updateUserStatus(int userId, Boolean updatedUserStatus);
+    public void updateUserStatus(long userId, Boolean updatedUserStatus);
 
 
     // update profile
